backend/requirements.txt
fastapi
uvicorn
google-generativeai
reportlab
python-multipart
